requirements.txt
anyio==4.6.0
asgiref==3.8.1
certifi==2024.8.30
chapa==0.1.2
Django==5.1.1
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
pillow==10.4.0
PyJWT==2.9.0
pytz==2024.2
sniffio==1.3.1
sqlparse==0.5.1
tzdata==2024.2
